+++ kernel/syscall.h
@@ -59,7 +59,7 @@
 
 __cdecl int
 syscall_handle(
-    uint32_t a, uint32_t b, uint32_t c, uint32_t d, uint32_t e,
+    intptr_t a, intptr_t b, intptr_t c, intptr_t d, intptr_t e,
     int_regs_t *regs, int num);
 
 #endif /* ASM */
